changed DB structure

+++ db/migrate/20151103003058_create_items.rb
@@ -3,7 +3,7 @@ class CreateItems < ActiveRecord::Migration
     create_table :items do |t|
       t.string :name
       t.string :description
-      t.integer :unit_price
+      t.float :unit_price
       t.integer :merchant_id
 
       t.timestamps null: false
